General: refactor modules, profiles, and programs
Just found out about home-manager.sharedModules. I've been searching
for these for HOURS now, and it turns out it's in Appendix B this
whole time?! My mind was blown - I'm not sure if I should feel sad
that I've wasted this much time (and almost opened an issue) on this,
or ecstatically happy that I've finally found a solution. I guess both
will have to do it.